Prefixed .bash_prompt's functions with prompt_
Refactored .bash_prompt functions to contain the prefix `prompt_`.

This is aimed to increase cosistency throughout the project, along with
all of the variables that already contain the prefix. This will also
help reduce conflicts with shell autocompletion, since all functions
are being exported to the environment and can be invoked from an
interactive shell.